♻️ 按语言拆分 Monaco 编辑器语言包为独立模块
langs.ts 已含 9 种语言约 1300 行，拆为 langs/<locale>.ts + index.ts，
导入路径 ./langs 不变。拆分为纯机械搬移，已用临时深度对比测试验证
拆分前后 editorLangs 内容完全一致。
