+++ src/pkg/utils/monaco-editor/langs/ru-RU.ts
@@ -1,0 +1,128 @@
+const grantValuePrompts = {
+  none: "Не запрашивает специальные права GM API; скрипт работает ближе к обычному скрипту страницы.",
+  unsafeWindow: "Доступ к собственному объекту window страницы для взаимодействия с нативными скриптами страницы.",
+  GM_getValue: "Читает одно значение из постоянного хранилища скрипта.",
+  GM_getValues: "Читает несколько значений из постоянного хранилища скрипта.",
+  GM_setValue: "Записывает одно значение в постоянное хранилище скрипта.",
+  GM_setValues: "Записывает несколько значений в постоянное хранилище скрипта.",
+  GM_deleteValue: "Удаляет одно значение из постоянного хранилища скрипта.",
+  GM_deleteValues: "Удаляет несколько значений из постоянного хранилища скрипта.",
+  GM_listValues: "Перечисляет все ключи в постоянном хранилище скрипта.",
+  GM_addValueChangeListener: "Отслеживает изменения значений в хранилище скрипта.",
+  GM_removeValueChangeListener: "Удаляет слушатель изменений значений в хранилище скрипта.",
+  GM_xmlhttpRequest: "Выполняет cross-origin сетевые запросы; целевые хосты обычно нужно разрешить через @connect.",
+  GM_download:
+    "Загружает файлы. Принимает URL и имя файла либо объект параметров с полями url, name, headers, saveAs и возвращает дескриптор с abort.",
+  GM_openInTab: "Открывает новую вкладку с параметрами, например в фоне или на переднем плане.",
+  GM_closeInTab: "Закрывает вкладку, открытую или управляемую скриптом.",
+  GM_getTab: "Читает временные данные, связанные с текущей вкладкой.",
+  GM_saveTab: "Сохраняет временные данные, связанные с текущей вкладкой.",
+  GM_getTabs: "Читает все временные данные вкладок, сохраненные скриптом.",
+  GM_notification: "Показывает уведомление браузера и обрабатывает события, например клик или закрытие.",
+  GM_closeNotification: "Закрывает указанное уведомление скрипта.",
+  GM_updateNotification: "Обновляет указанное уведомление скрипта.",
+  GM_setClipboard: "Записывает данные в системный буфер обмена.",
+  GM_registerMenuCommand: "Регистрирует команду меню скрипта.",
+  GM_unregisterMenuCommand: "Отменяет регистрацию команды меню скрипта.",
+  CAT_registerMenuInput: "API ScriptCat: регистрирует команду меню скрипта с полем ввода.",
+  CAT_unregisterMenuInput: "API ScriptCat: отменяет регистрацию команды меню скрипта с полем ввода.",
+  GM_addStyle: "Внедряет CSS-стили на страницу.",
+  GM_addElement: "Создает и вставляет элемент на страницу.",
+  GM_getResourceText: "Читает текстовое содержимое ресурса, объявленного через @resource.",
+  GM_getResourceURL: "Получает URL ресурса, объявленного через @resource.",
+  GM_cookie: "Доступ к Cookie API для чтения, записи или удаления cookie.",
+  CAT_fetchBlob: "Внутренний API ScriptCat: читает доступный со стороны расширения ресурс и возвращает Blob.",
+  CAT_fileStorage: "API ScriptCat: доступ к файловому хранилищу скрипта.",
+  CAT_userConfig: "API ScriptCat: доступ к пользовательской конфигурации скрипта.",
+  CAT_scriptLoaded: "API ScriptCat: ожидает полной загрузки скрипта в сценариях @early-start.",
+  "window.close": "Разрешает скрипту вызывать window.close().",
+  "window.focus": "Разрешает скрипту вызывать window.focus().",
+  "window.onurlchange": "Разрешает скрипту слушать события изменения URL.",
+} as const;
+
+export default {
+  title: "Русский",
+  thisIsAUserScript: "Пользовательский скрипт",
+  undefinedPrompt: "Неопределённый промпт",
+  quickfix: "Исправить проблему {0}",
+  addEslintDisableNextLine: "Добавить комментарий eslint-disable-next-line",
+  addEslintDisable: "Добавить комментарий eslint-disable",
+  declareGlobal: "Объявить '{0}' как глобальную переменную (/* global */)",
+  removeConnectWildcard: "Удалить wildcard @connect: {0}",
+  replaceMatchTldWildcardWithInclude: "Заменить TLD wildcard @match на @include {0}",
+  replaceIncludeWithMatch: "Заменить @include на @match {0}",
+  grantConflict: "@grant none нельзя использовать вместе с GM API. Удалите none или все GM API.",
+  grantValuePrompts,
+  prompt: {
+    name: "Имя скрипта",
+    namespace: "Пространство имён скрипта",
+    copyright: "Информация об авторских правах скрипта",
+    license: "Лицензия с открытым исходным кодом",
+    version: "Версия скрипта",
+    description: "Описание скрипта",
+    icon: "Иконка скрипта",
+    iconURL: "Иконка скрипта",
+    defaulticon: "Иконка скрипта",
+    icon64: "Иконка скрипта 64x64",
+    icon64URL: "Иконка скрипта 64x64",
+    grant: "Запрашиваемые специальные права доступа к API",
+    author: "Автор скрипта",
+    "run-at":
+      "Момент запуска скрипта<br>`document-start`: внедрить как можно раньше после совпадения URL<br>`document-end`: внедрить после загрузки DOM (изображения и др. могут ещё загружаться)<br>`document-idle`: внедрить после полной загрузки содержимого<br>`document-body`: внедрить только если на странице есть элемент body",
+    "run-in": "Контекст, в который внедряется скрипт",
+    homepage: "Домашняя страница скрипта",
+    homepageURL: "Домашняя страница скрипта",
+    website: "Домашняя страница скрипта",
+    background: "Фоновый скрипт",
+    include: "Страницы, на которых скрипт выполняется (совпадение URL)",
+    match: "Страницы, на которых скрипт выполняется (совпадение URL)",
+    exclude: "Страницы, на которых скрипт НЕ выполняется (совпадение URL)",
+    connect: "Сайты, к которым скрипт может обращаться",
+    resource: "Подключаемые ресурсные файлы",
+    require: "Подключаемые внешние JS-файлы",
+    "require-css": "Подключаемые внешние CSS-файлы",
+    noframes: "Не запускать скрипт внутри `<frame>`",
+    compatible: "Информация о совместимости, отображаемая на GreasyFork",
+    "inject-into":
+      "Контекст внедрения скрипта<br>`content`: внедрить в контекст content<br>`page`: внедрить в контекст страницы (по умолчанию)<br>Примечание: SC не поддерживает `inject-into: auto`, когда контекст выбирается по CSP.",
+    "early-start":
+      "Используется совместно с `run-at: document-start`. `early-start` позволяет выполнять скрипт раньше загрузки страницы, но может ухудшать производительность и ограничивать некоторые GM API. (Только в SC)",
+    unwrap:
+      "Позволяет пользовательскому скрипту обходить песочницу и напрямую внедряться и выполняться в нативной глобальной области видимости страницы. <br>Скрипт может напрямую получать доступ к реальным глобальным переменным страницы и изменять их, однако не сможет использовать привилегированные API пользовательских скриптов, такие как GM.*. <br>Обычно используется в сценариях, требующих глубокой интеграции с нативными скриптами страницы или при миграции с обычных скриптов страницы.",
+    definition: "Особенность ScriptCat: URL файла `.d.ts`, используемого для автодополнения в редакторе",
+    antifeature: `Связано с маркетплейсами скриптов: для нежелательных функций следует добавить это значение описания
+referral-link: Этот скрипт изменяет или перенаправляет на реферальную ссылку автора
+ads: Этот скрипт вставляет рекламу на посещаемые вами страницы
+payment: Этот скрипт требует оплаты для нормального использования
+miner: Этот скрипт содержит функции майнинга
+membership: Этот скрипт требует регистрации членства для нормального использования
+tracking: Этот скрипт отслеживает информацию о пользователе`.replace(/\n/g, "<br>"),
+    updateURL: "URL для проверки обновлений скрипта",
+    downloadURL: "URL для загрузки обновлений скрипта",
+    supportURL: "Страница поддержки / отчёта об ошибках",
+    source: "Страница с исходным кодом скрипта",
+    scriptUrl: "URL пользовательского скрипта, на который ссылается скрипт подписки",
+    storageName: "Имя хранилища значений скрипта для совместного использования одного хранилища несколькими скриптами",
+    tag: "Теги скрипта, разделённые запятыми или пробелами",
+    cloudCat: "Отмечает, что скрипт можно экспортировать в пакет облачного скрипта CloudCat",
+    cloudServer: "Облачный сервис CloudCat, используемый скриптом",
+    exportValue: "Значения хранилища скрипта для экспорта при экспорте как облачного скрипта",
+    exportCookie: "Cookie для экспорта при экспорте как облачного скрипта",
+    crontab: `Примеры crontab для планового запуска скриптов (не для облачных скриптов)
+* * * * * * Запуск каждую секунду
+* * * * * Запуск каждую минуту
+0 */6 * * * Запуск раз в 6 часов в 00 минут
+15 */6 * * * Запуск раз в 6 часов в 15 минут
+* once * * * Запуск раз в час
+* * once * * Запуск раз в день
+* 10 once * * Запуск раз в день между 10:00-10:59; если выполнен в 10:04, в этот день не запустится снова между 10:05-10:59
+* 1,3,5 once * * Запуск раз в день в 1:00, 3:00, 5:00; если выполнен в 1:00, в этот день не запустится в 3:00 и 5:00
+* */4 once * * Проверка и запуск раз в 4 часа; если выполнен в 4:00, в этот день не запустится в 8:00, 12:00, 16:00, 20:00, 24:00
+* 10-23 once * * Запуск раз в день между 10:00-23:59; если выполнен в 10:04, в этот день не запустится снова между 10:05-23:59
+* once 13 * * Запуск каждый час в течение 13-го числа месяца
+* once(9-17) * * * Каждый день с 9 до 17 часов, запуск раз в час
+0,30 once * * * Раз в час в минуту 0 или 30, первое срабатывание побеждает, второе пропускается
+* * once(9-18) * * С 9 по 18 число каждого месяца, запуск раз в день
+* * * * once(1-5) Раз в неделю, только с понедельника по пятницу`.replace(/\n/g, "<br>"),
+  },
+} as const;
